requirements.txt
anyio==4.0.0
black==23.10.0
cached-property==1.5.2
certifi==2023.7.22
click==8.1.7
exceptiongroup==1.1.3
filelock==3.12.4
fsspec==2023.9.2
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
inflection==0.5.1
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.26.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
psycopg2-binary==2.9.9
python-telegram-bot==20.6
pytz==2023.3.post1
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlalchemy-orm==1.2.10
sympy==1.12
tomli==2.0.1
torch==2.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
